test(query): move test case
